+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.2.1 (unreleased)
 
+IMPROVEMENTS:
+
+  * core: Plugins are automatically discovered in the executable directory
+      or pwd if named properly. [GH-190]
+
 BUG FIXES:
 
   * core: Configuration parses when identifier and '=' have no space. [GH-243]